add tests for acces rules module
