+++ tests/integration/room-access-rules/public-room.spec.ts
@@ -1,39 +1,10 @@
 import { test, expect } from '@playwright/test';
 import type { MatrixApi } from '../../../utils/matrix-api';
 import {  deactivateMasUser } from '../../../utils/mas-admin';
-import { expectErrorWhenSendStateEvent, loginWithNewUser } from './room-utils';
+import { createPublicRoom, expectErrorWhenSendStateEvent, loginWithNewUser } from './room-utils';
 import { EventType } from 'matrix-js-sdk';
 
-export async function createPublicRoom(
-  matrix: MatrixApi,
-  name: string = 'Public Room'
-): Promise<string> {
-  return matrix.createRoom({
-    name,
-    joinRule: 'public',
-    preset: 'public_chat',
-    visibility: 'public',
-    accessRules: {
-      rule: 'restricted',
-      force_unencrypted_at_creation: false,
-      visibility: 'public',
-    },
-    power_level_content_override:{
-      events:
-      {
-         "m.room.name": 50,
-          "m.room.avatar": 50,
-          "m.room.power_levels": 100,
-          "m.room.history_visibility": 100,
-          "m.room.canonical_alias": 50,
-          "m.room.tombstone": 100,
-          "m.room.server_acl": 100,
-          "m.room.encryption": 100,
-          "org.matrix.msc3401.call.member": 0
-      },
-    }
-  });
-}
+
 
 test.describe('API - Public Room', () => {
   let userId: string;
