+++ tests/integration/room-access-rules/room-utils.ts
@@ -35,12 +35,9 @@ export async function expectErrorWhenSendStateEvent(
   content: Record<string, any>,
   expectedStatus: number = 403
 ): Promise<void> {
-  try {
-    await matrix.sendStateEvent(roomId, eventType, content);
-    throw new Error(`Expected ${expectedStatus} error but request succeeded`);
-  } catch (error: any) {
-    expect(error.httpStatus).toBe(expectedStatus);
-  }
+  await expect(matrix.sendStateEvent(roomId, eventType, content))
+    .rejects
+    .toMatchObject({ httpStatus: expectedStatus });
 }
 
 export async function createPrivateEncryptedRoom(
@@ -57,5 +54,55 @@ export async function createPrivateEncryptedRoom(
       force_unencrypted_at_creation: false,
       visibility: 'private',
     },
+  });
+}
+
+export async function createPrivateUnencryptedRoom(
+  matrix: MatrixApi,
+  name: string = 'Private Unencrypted Room'
+): Promise<string> {
+  return matrix.createRoom({
+    name,
+    joinRule: 'invite',
+    preset: 'private_chat',
+    visibility: 'private',
+    encryption: false,
+    accessRules: {
+      rule: 'restricted',
+      force_unencrypted_at_creation: true,
+      visibility: 'private',
+    },
+  });
+}
+
+
+export async function createPublicRoom(
+  matrix: MatrixApi,
+  name: string = 'Public Room'
+): Promise<string> {
+  return matrix.createRoom({
+    name,
+    joinRule: 'public',
+    preset: 'public_chat',
+    visibility: 'public',
+    accessRules: {
+      rule: 'restricted',
+      force_unencrypted_at_creation: false,
+      visibility: 'public',
+    },
+    power_level_content_override:{
+      events:
+      {
+         "m.room.name": 50,
+          "m.room.avatar": 50,
+          "m.room.power_levels": 100,
+          "m.room.history_visibility": 100,
+          "m.room.canonical_alias": 50,
+          "m.room.tombstone": 100,
+          "m.room.server_acl": 100,
+          "m.room.encryption": 100,
+          "org.matrix.msc3401.call.member": 0
+      },
+    }
   });
 }
